+++ src/Products/PlonePAS/browser/info.py
@@ -1,12 +1,11 @@
 # -*- coding: utf-8 -*-
 from Acquisition import aq_inner
+from plone.memoize.instance import memoize
 from Products.CMFCore.utils import getToolByName
 from Products.Five import BrowserView
 from Products.PlonePAS.interfaces.browser import IPASInfoView
 from Products.PluggableAuthService.interfaces.plugins import IExtractionPlugin
-from Products.PluggableAuthService.interfaces.plugins \
-    import ILoginPasswordExtractionPlugin
-from plone.memoize.instance import memoize
+from Products.PluggableAuthService.interfaces.plugins import ILoginPasswordExtractionPlugin  # noqa: E501
 from zope.interface import implementer