+++ src/Products/PlonePAS/tests/test_memberdatatool.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 from DateTime import DateTime
 from OFS.Image import Image
+from plone.app.testing import TEST_USER_ID as default_user
 from Products.PlonePAS.tests import base
 from Products.PlonePAS.tests import dummy
-from plone.app.testing import TEST_USER_ID as default_user
 
 
 class TestMemberDataTool(base.TestCase):